Move modal to new component and add static input fields
