+++ editor.planx.uk/src/pages/FlowEditor/components/Team/TeamMembers.tsx
@@ -1,14 +1,10 @@
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
 import Container from "@mui/material/Container";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
 import Typography from "@mui/material/Typography";
 import React, { useState } from "react";
 import SettingsSection from "ui/editor/SettingsSection";
 
-import { MembersTable } from "./MembersTable";
+import { AddNewEditorModal } from "./components/AddNewEditorModal";
+import { MembersTable } from "./components/MembersTable";
 import { TeamMember, TeamMembersProps } from "./types";
 
 export const TeamMembers = ({ teamMembersByRole }: TeamMembersProps) => {
@@ -66,57 +62,7 @@ export const TeamMembers = ({ teamMembersByRole }: TeamMembersProps) => {
         </SettingsSection>
       )}
       {showModal && (
-        <Dialog
-          // maxWidth="xl"
-          // aria-labelledby="dialog-heading"
-          PaperProps={{
-            sx: {
-              width: "100%",
-              maxWidth: (theme) => theme.breakpoints.values.md,
-              borderRadius: 0,
-              borderTop: (theme) => `20px solid ${theme.palette.primary.main}`,
-              background: "#FFF",
-              margin: (theme) => theme.spacing(2),
-            },
-          }}
-          open={showModal}
-          onClose={() => setShowModal(false)}
-        >
-          <DialogContent>
-            <Box sx={{ mt: 1, mb: 4 }}>
-              <Typography variant="h3" component="h2" id="dialog-heading">
-                Add a new editor
-              </Typography>
-            </Box>
-          </DialogContent>
-          <DialogActions
-            sx={{
-              display: "flex",
-              justifyContent: "flex-start",
-              padding: 2,
-            }}
-          >
-            <Box>
-              <Button
-                variant="contained"
-                color="prompt"
-                onClick={() => setShowModal(false)} // nothing yet
-                data-testid="modal-create-user-button"
-              >
-                Create user
-              </Button>
-              <Button
-                variant="contained"
-                color="secondary"
-                sx={{ ml: 1.5 }}
-                onClick={() => setShowModal(false)}
-                data-testid="modal-cancel-button"
-              >
-                Cancel
-              </Button>
-            </Box>
-          </DialogActions>
-        </Dialog>
+        <AddNewEditorModal showModal={showModal} setShowModal={setShowModal} />
       )}
     </Container>
   );
